+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [0.1.43]
+* Update GetServicePrincipalToken to allow using certificate for App Reg authentication
+* Update deps
+
 ## [0.1.42]
 * Bugfix: Fix to handling lib.AzureMultiAuthTokenRequestOptions.NoCache